+++ src/Rector/Annotation/AnnotationReplacerRector.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Rector\Dynamic;
+namespace Rector\Rector\Annotation;
 
 use PhpParser\Node;
 use PhpParser\Node\Stmt\ClassMethod;
