+++ tests/Rector/Annotation/AnnotationReplacerRector/Correct/correct.php.inc
