+++ tests/Rector/Constant/ClassConstantReplacerRector/Correct/correct.php.inc
