+++ tests/Rector/Constant/ClassConstantReplacerRector/Wrong/wrong.php.inc
