+++ tests/Rector/MethodCall/MethodNameReplacerRector/config.yml
@@ -1,5 +1,5 @@
 services:
-    Rector\Rector\Dynamic\MethodNameReplacerRector:
+    Rector\Rector\MethodCall\MethodNameReplacerRector:
         $perClassOldToNewMethods:
             'Symfony\Component\Form\AbstractType':
                   'setDefaultOptions': 'configureOptions'
