+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/SomeClass.php
