+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/correct.php.inc
