+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/correct2.php.inc
