+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/correct3.php.inc
