+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/correct4.php.inc
