+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/correct5.php.inc
