+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/correct6.php.inc
