+++ tests/Rector/MethodCall/MethodNameReplacerRector/correct/correct7.php.inc
