+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/SomeClass.php
