+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/wrong.php.inc
