+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/wrong2.php.inc
