+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/wrong3.php.inc
