+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/wrong4.php.inc
