+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/wrong5.php.inc
