+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/wrong6.php.inc
