+++ tests/Rector/MethodCall/MethodNameReplacerRector/wrong/wrong7.php.inc
