+++ tests/Rector/Namespace_/NamespaceReplacerRector/Correct/correct.php.inc
