+++ tests/Rector/Namespace_/NamespaceReplacerRector/Correct/correct2.php.inc
