+++ tests/Rector/Namespace_/NamespaceReplacerRector/Correct/correct3.php.inc
