+++ tests/Rector/Namespace_/NamespaceReplacerRector/Correct/correct4.php.inc
