+++ tests/Rector/Namespace_/NamespaceReplacerRector/Correct/correct5.php.inc
