+++ tests/Rector/Namespace_/NamespaceReplacerRector/Wrong/wrong.php.inc
