+++ tests/Rector/Namespace_/NamespaceReplacerRector/Wrong/wrong2.php.inc
