+++ tests/Rector/Namespace_/NamespaceReplacerRector/Wrong/wrong3.php.inc
