+++ tests/Rector/Namespace_/NamespaceReplacerRector/Wrong/wrong4.php.inc
