+++ tests/Rector/Namespace_/NamespaceReplacerRector/Wrong/wrong5.php.inc
