+++ tests/Rector/Namespace_/NamespaceReplacerRector/config.yml
@@ -1,5 +1,5 @@
 services:
-    Rector\Rector\Dynamic\NamespaceReplacerRector:
+    Rector\Rector\Namespace_\NamespaceReplacerRector:
         $oldToNewNamespaces:
             'OldNamespace': 'NewNamespace'
             'OldNamespaceWith\OldSplitNamespace': 'NewNamespaceWith\NewSplitNamespace'
