+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Correct/correct.php.inc
