+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Correct/correct2.php.inc
