+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Correct/correct3-expected.php.inc
