+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Correct/correct3.php.inc
