+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Correct/correct4.php.inc
