+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/PseudoNamespaceToNamespaceRectorTest.php
@@ -1,11 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Dynamic\ClassConstantReplacerRector;
+namespace Rector\Tests\Rector\Namespace_\PseudoNamespaceToNamespaceRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-final class ClassConstantReplacerRectorTest extends AbstractRectorTestCase
+final class PseudoNamespaceToNamespaceRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideWrongToFixedFiles()
@@ -18,6 +18,9 @@ final class ClassConstantReplacerRectorTest extends AbstractRectorTestCase
     public function provideWrongToFixedFiles(): Iterator
     {
         yield [__DIR__ . '/Wrong/wrong.php.inc', __DIR__ . '/Correct/correct.php.inc'];
+        yield [__DIR__ . '/Wrong/wrong2.php.inc', __DIR__ . '/Correct/correct2.php.inc'];
+        yield [__DIR__ . '/Wrong/wrong3.php.inc', __DIR__ . '/Correct/correct3.php.inc'];
+        yield [__DIR__ . '/Wrong/wrong4.php.inc', __DIR__ . '/Correct/correct4.php.inc'];
     }
 
     protected function provideConfig(): string
