+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Wrong/wrong.php.inc
