+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Wrong/wrong2.php.inc
