+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Wrong/wrong3.php.inc
