+++ tests/Rector/Namespace_/PseudoNamespaceToNamespaceRector/Wrong/wrong4.php.inc
